# Enumerable.Index/Select((T, int)) iterator classes