@Emik03 2024-05-16
Oh dear, the tables turned out far less readable than it looked in preview, I don't know if I can do anything to make them look better.