@stephentoub 2024-05-16
> Binary Size

What binary size is being measured here, the System.Linq.dll assembly?  While that's interesting, the more impactful size is assembly, e.g. if you publish a NativeAOT app that's using LINQ, say a Select operator somewhere, what does the addition of these types do to that NativeAOT published binary size?