@Emik03 2024-05-17
> > Binary Size
> 
> What binary size is being measured here, the System.Linq.dll assembly? While that's interesting, the more impactful size is assembly, e.g. if you publish a NativeAOT app that's using LINQ, say a Select operator somewhere, what does the addition of these types do to that NativeAOT published binary size?

I originally thought file-size concerns were regarding `System.Linq.dll`, my mistake. I'll edit the post tomorrow to additionally include a basic NativeAOT test, with the source code I'm testing with.