@eiriktsarpalis 2024-05-17
In light of the inconclusive results, I think it probably isn't worth applying iterator optimizations to `Index`. Primarily the method is intended to faciliate foreach expressions:
```C#
foreach ((int i, T item) in source.Index) { }
```
And I wouldn't consider it common enough for people to chain it with `ToArray`/`ToList`/`TryGetFirst` calls.