@Emik03 2024-05-17
I've been trying all day but had trouble compiling NativeAOT with the self-built `System.Linq`, so I can't tell you how this fares in size.

However, my idea is that if performance and binary size stays more-or-less the same, isn't that already good enough? Perhaps the average use-case isn't optimized, but if `.Select` or `.Index` ever gets passed directly or indirectly into `.ToArray()` or similar, then having the count be computable is a small improvement.