@eiriktsarpalis 2024-05-17
> However, my idea is that if performance and binary size stays more-or-less the same, isn't that already good enough?

It also replaces a simple implementation with a more complex one. Any change comes with added cost, including risk of potential regression so this would need to be outweighed with measurable benefits.