@eiriktsarpalis 2024-05-17
That being said, I really appreciate your effort and the attention to detail that you put into this PR! I think we collectively learned something by it, so let this not dissuade you from similar experimentation in the future.