@Emik03 2024-05-17
> > However, my idea is that if performance and binary size stays more-or-less the same, isn't that already good enough?
> 
> It also replaces a simple implementation with a more complex one. Any change comes with added cost, including risk of potential regression so this would need to be outweighed with measurable benefits.

If I may ask, what kind of potential regressions could you be referring to? All tests pass just fine, and performance is the same, or slightly better for `Index` (speed variant) in lists/arrays. I hope I haven't repeated myself too much by saying that, I'm just having trouble understanding the potential problems here.