@eiriktsarpalis 2024-05-17
> If I may ask, what kind of potential regressions could you be referring to?

No specific concerns, I'm just pointing out that every PR that we merge introduces risk and maintenance cost. If it doesn't move the needle substantially, we err towards not making the change at all.