@stephentoub 2024-06-23
Closing per https://github.com/dotnet/runtime/pull/102314#issuecomment-2117546591. But thanks!